perf(web): defer tour overlay + lazy-load sidebar animation engine (F9)
Extracts tour.tsx's motion-heavy spotlight/cursor/content overlay into
tour-overlay.tsx, loaded via next/dynamic(ssr:false) — most /gen
visitors never start the tour, so the chunk (and the animation engine
it pulls in) is only fetched once a tour actually starts, not on every
page load. TourProvider latches an `everStarted` flag so the overlay
stays mounted after first use, letting its own AnimatePresence track
exit animations across steps and on close correctly.

Migrates sidebar.tsx from motion.div/motion.span to LazyMotion + m.div
(domMax, since the active-nav-item indicator needs layoutId shared-
layout animation) — splits the animation engine into a separately
fetched chunk instead of inlining it with every file that imports
motion/react.

hero-entrance.tsx/hero-showcase.tsx/animated-header.tsx are left eager:
they render on first paint of the routes that use them (the header via
every page's SiteShell) — there's no later point to defer them to.
